+++ test/ZfcUserSettingsTest/Mapper/DoctrineORM/UserSettingMapperTest.php
@@ -4,6 +4,7 @@ namespace Eye4web\ZfcUserSettingsTest\Doctrine\ORM;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Eye4web\ZfcUser\Settings\Mapper\DoctrineORM\UserSettingMapper;
+use Eye4web\ZfcUser\Settings\Entity\Setting;
 use PHPUnit_Framework_TestCase;
 use ZfcUser\Entity\User;
 
@@ -30,16 +31,11 @@ class UserSettingMapperMapperTest extends PHPUnit_Framework_TestCase
 
     public function testGetSetting()
     {
-     /*   $objectRepository = $this->getMock('Doctrine\Common\Persistence\ObjectRepository');
+        $setting = new Setting();
         $this->objectManager->expects($this->once())
-                            ->method('getRepository')
-                            ->with('Eye4web\ZfcUser\Settings\Entity\Setting')
-                            ->will($this->returnValue($objectRepository));
-        $objectRepository->expects($this->once())
-                   ->method('find')
-                   ->will($this->returnValue(new Setting()));
-
-        $this->assertInstanceOf('ZfcUser\Settings\Entity\Setting', $this->mapper->getSetting(1));
-      */
+                            ->method('find')
+                            ->with('Eye4web\ZfcUser\Settings\Entity\Setting', 1)
+                            ->will($this->returnValue($setting));
+        $this->assertSame($setting, $this->mapper->getSetting(1));
     }
 }